Gives some enemies guaranteed drops on their guns (#4608)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Some ruins relied on non-neutered enemies having a 100% drop chance for
certain loot drops. This got removed when a mostly-global drop chance
was applied to simplehumans making it significantly easier to have 20 of
them only maybe drop 1 gun
Specifically:
The frontie commander in distillery now drops their mauler 100% of the
time so the magazines on the table are useful
The frontie shotgunner in abandoned village (remember that one? Me
neither) now drops their shotgun 100% of the time so the slugs in the
safe aren't just paperweights
The manager in wrecked factory now uses the hades that was in the
display case, instead of having a second one they manifested out of
sheer hatred for ramzi.

This is a COOL TRICK you can use to turn an enemy into a loot drop at
the same time

## Why It's Good For The Game

My loot.....

## Changelog

:cl:
balance: certain enemies at the end of ruins are now significantly more
likely to drop their guns
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

(cherry picked from commit 41c4b0d224d1e6d3629a58d698af6b63a52d3245)