fixes wielding (#4643)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Adds a flag since the target of the wield channel is ALSO the user so
they have to have BOTH ignore user loc and ignore target loc flags to
not fail if you move. Also puts it back where it belongs since it's set
to be FALSE only to be directly told to actually not do that at the end
for some reason

## Why It's Good For The Game

uhm. yeah.

## Changelog

:cl:
fix: the wield channel will no longer fail if you move
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

(cherry picked from commit 9febb3d470de1c9cff726c6c7a76823a74a6ab27)