Indie deckhand outfit tweaks (#4598)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
It always irked me that every deckhand spawns with `A somewhat
uncomfortable suit designed to be as cheap as possible to manufacture.`

Now, instead, most indies start with a much more neutral black jumpsuit.
The utility jumpsuit is, however, preserved for a `/cheap` variant that,
similarly to how the Captain outfit does it, is to be used on Miskilamo
ships.
Also, makes the Corona use the `/fancy` outfit added with the Ivory.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Changelog

:cl:
add: Deckhands on non-Miskilamo ships now start with a black jumpsuit
instead of a utility one. Corona deckhands also got fancier.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

(cherry picked from commit a585f7ecb2164f30229fc68d3421300b172a9d68)