Bone breaking rework (#4568)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Fracturing a body part and the likelihood of it now depends on the
actual force of the attack, becoming easier and more likely to fracture
the more your body part is damaged.
This does mean that you can actually break a bone from full health, but
right now that requires receiving 50+ brute damage in a single attack,
which is exceedingly rare.
This also means that you can no longer break bones from something like
decompression, hail, sand storms etc., since it deals less than 5 damage
per individual proc call.

Force check for being able to fracture:
- Damage taken >= (Damage Threshold - clamp((Existing damage * 0.5), 0,
Damage Threshold - 5)
- Where Damage Threshold = 40 (45 for head, 25 for everything else)

Formula for percentage chance of fracture:
- Raw Force of Weapon/Projectile + Existing Damage * 0.5

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Fractures have felt a little unintuitive, mechanically shallow and
nonsensical. This aims to solve that through tying the ability to
fracture and the probability of inflicting a fracture to the received
damage, existing damage and force of the attack.
This also gives armor more importance/usefulness, since it decreases
damage taken, which increases the raw attack force required to create a
fracture.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
balance: Fractures now rely more on received damage and the force of the
attack.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: DrCrawler <62419802+DrCrawler@users.noreply.github.com>
Co-authored-by: Theos <theubernyan@gmail.com>
(cherry picked from commit 0179e29f51c9ebb07cf26f92624fc1d6d944df49)