Smarter Test (#314)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

This PR creates a redundancy backup system for the code to acquire byond
in cases of server downtime or DDOS.


Step 1: Download Byond
- Will attempt to connect to byond.com and download the version of Linux
and Windows builds as dictated by dependencies.sh.
- Time before failing to download has been extended, test will try for a
short while which could result in getting a direct copy.
- IF: Server is available: Tests will proceed like normal
- IF: Server is offline: Step 2

Step 2: Local Redundancy Check
- Will attempt to look for the same version, as dictated by
dependencies.sh inside .../tools/backup
- IF: File is located: Tests will proceed like normal
- IF: File is not located: Step 3

Step 3: Emergency Backup Version
- This check will (currently) always work.
- The test will look for the current 516.1663 file
- Tests will proceed like normal
- Should this fail, then the test will fail.


The reason I have coded the test in this way is because we should always
be testing in the environment we currently using. During this DDOS
downtime I was able to acquire copies of the windows builds from
515.1642 thru 516.1663 but not the linux builds.

This means that this test, during the downtime, is going to be testing
the 515.1642 version of windows, and (unless it can get the 515.1642
version from BYOND) the 516.1663 version of linux.

This means that, there is the possibility of issues with a PR if has
issues with the 516 version of Byond. Yet... Since most 516 features
appear to mostly be UI and player facing... this might not become a huge
issue.



What does this mean for future me (or others who do my job)? This means
that over time, whenever we update what version of the code we plan on
running with our dependencies.sh we should also download and include
that version inside the repo. The files themselves are small, but
honestly in a perfect world, we shouldn't have to do this.


Also the cache system has also been ported over to the windows build, to
help reduce the amount of hits to BYOND.



<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

I can actually get testing done without having to hope for RNG that the
server is not being DDOSed

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->