Refactor Fabricator Topic + Directional Printing (#283)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Refactors the mecha fabricator Topic proc to properly use switch, adds
directional printing, and fixes issues with printing into dense objects.
It was previously set to print in the dir of src, which was always
south. Now you can select the output direction from the menu, and if it
detects a problem, it will print on its own turf.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

This should make the fabricator menu snappier as it's not using a bunch
if checks for no reason. Plus it fixes my biggest issue with it, which
is printing in a single direction. That and a sprinkle of bug fixing.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Mech fabricator can now print in all cardinal directions,
selectable from the menu.
fix: The fabricator will not print into tiles with dense
objects/turfs/mobs, fallback to its own tile.
refactor: The fabricator menu should be snappier with switch().
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: Ossa88 (SYNAPSE) <MegaMan101@gmail.com>