IPC Integrity Be-Gone (#298)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

So lets be frank here,

This was a rather controversial change, and one I feel like is good for
the balance of IPCs...

BUT

Some issues have risen where the mechanic just started to become
bothersome and had to work around this entire mechanic. Primarily
involved making it so that wire based healing would no longer touch
integrity, because when an IPC enters soft critical they start to take
burn damage. Wire heals burns, add to integrity, crit keeps adding
burns, and so on until so much integrity damage is added that it becomes
near impossible to recover the IPC player.

What this left was only welder healing for brute damage touches
integrity, and as some have pointed out, this can be gotten around...

So I made the administrative choice to just disable it entirely, for
now, until a better idea can come out of the entire thing, and or a
newer mechanic can be made upstream... I currently have some idea's but
do not have the time to code them at this moment.

What does this mean? Structure kits will no longer have any valid use,
same for the structural surgery (both of which will still exist for the
time being), but the repair surgery will still be there for fixing dead
IPC friends and stuff.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Since only welding deals integrity loss to a body part, it means that
one entire damage source no longer an issue, and at that point why keep
this around, just makes things complicated.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
del: Robot Limbs and IPCs no longer have to deal with Integrity. Pending
future updates.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->