Ships Only Tile Within Own Area (#285)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

This fixes an issue where ships would tile with the environment/other
ships. There was sort of an exception for this in the code but it was
nonfunctional. I refactored the code just a tad and implemented a proper
method for checking this.

There is a caveat here that anything built outside of the ship will not
immediately tile with it, like when you are trying to expand it.
However, I've altered the ship blueprints to smooth the area. Simply
redefine your expansion/new area, and it will tile with the rest of the
ship.

Example 1:

![dreamseeker_Q9w9VXtKdC](https://github.com/user-attachments/assets/88a7083c-c103-46d6-9766-eebcf725ea0f)

Example 2:

![dreamseeker_szl92y2edJ](https://github.com/user-attachments/assets/3f11aaf7-c494-489d-9305-ba9bc76f060d)

Video (on Discord since the file was too big):

https://discord.com/channels/1273549503030493204/1273575419446169631/1366240706980483092

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Are you tired of your ship dissolving into a formless blob of metal when
visiting any trash planet? So am I.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
fix: Ships will only tile within themselves and not the
environment/other ships. Using the shuttle expansion permit will
resmooth new/updated areas.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: Ossa88 (SYNAPSE) <MegaMan101@gmail.com>