i have written an extremely glorious read me.
if you work on books on this codebase and you don't read this readme, i'm going to be a bit upset. because it took me so long to write it. i mean, this was really the cumulative work of days, weeks even. i cobbled it together for you, to save you the minutes and hours i toiled just to understand what the fuck was going on. this was my sacrifice. this was my masterpiece. this was my